Tighten init template validation
